worker: set gdn flag defaults manually
Using go-flags to set Concourse specific defaults for gdn binary flags
overrides any value the user may pass via the config.ini file for
Guardian. These can be instead set manually when the values are not
present in the config and the defaults can still be stated in the
command description.

gdn specific flags have been moved to their own struct, GdnBinaryFlags,
to avoid confusion.

Following conditions are met in the order given
1. Sets GdnBinaryFlag when it has been set by user either via env or CLI option
2. Does nothing if flag is present in config.ini
3. Sets default value

There might be a simpler way to do this in the future using the go-flags
IniParser that Guardian uses to parse the config.ini.

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>