web/db/fly: behaviour: order pipelines by name
concourse/concourse#5921

rather than by `atc.PipelineRef`, which also includes the instance vars.
Also, when a pipeline is saved, if it has the same name as an existing
pipeline, use that existing pipelines ordering value as the new
ordering. With these two changes, all instances within an instance group
will have the same ordering. Ties are broken by pipeline ID, which means
pipelines created earlier appear first.

The reason we want all instances to have the same ordering is to ensure
cards don't move around on the dashboard when instances get
auto-archived, which would be a weird UX.

The plan is to not support drag and drop within an instance group in the
MVP, as that would imply the need for a secondary ordering within the
instance group.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>