web: behaviour: allow fuzzy search by instance var values
concourse/concourse#5921

This fuzzy search thing is pretty SWEET - instance var values can be
fuzzy matched in any order by just putting spaces between the search
terms. This doesn't match the initial proposal for searching - that will
come in a later version of Concourse when we implement proper structured
filtering of instance groups and use that to determine if we're in the
instance group view.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>