web: structure: fix watsjs
concourse/concourse#5921

The tests were using the old route `/teams/:team/pipelines/:pipeline`,
rather than `/pipelines/:id`. Since the ID cannot be determined
statically, I added a `setPipeline` helper that returns the ID of the
created pipeline.

Also, due to some changes to the lazy rendering, we have to resize the
screen so that 5 pipelines can be displayed at once.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>