testflight/topgun: move resource container gc test
the topgun test was a little too particular - not only did it check that
the container was GCed, it also checked that the resource config was
GCed.

resource configs live for a while longer now - up to an hour + 5
minutes. i don't think it's really important to test for this; the
user-facing impact is that the containers go away, they shouldn't care
about the database.

if we just test that the container goes away, it's super easy to cover
this in testflight tests, which are also much faster.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>