web: behavior: show check output on resource page
below the check status header, show the build output - typically just a
single check step.

a followup commit will change this to only show the status header if
you're not authorized to view the build output. this will look a little
more tidy. a bunch of tests will have to be modified, so let's commit
this first.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>