atc: behavior: name check builds 'check'
my name is check

giving them a one-off build name is kinda weird. in fact, one-off build
names are weird in general.

the name 'check' is at least slightly explanatory - it should be obvious
enough in 'fly builds' output for example.

a couple of followup commits will adjust 'fly builds' and the web UI (if
someone browses to a resource check build for whatever reason) to make a
bit more sense.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>