atc: structure: remove (exec.Step).Succeeded
instead of the .Succeeded getter method, which forces steps to be
stateful and has caused nil panics in the past, just return a bool from
Run indicating step's success or failure.

this commit updates the atc/engine/ tests, since but not atc/exec/ yet -
splitting that out into a second commit to make reviewing a little
easier.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>