atc: structure: add (exec.RunState).Run
allows a step to easily run a build plan dynamically. this has quite a
few applications:

* whether to run the 'get' after a 'put' can be determined at runtime
  rather than at build plan construction time, and given custom UI if
  desired
* task's image_resource fetching can be done by running a CheckPlan and
  a GetPlan, allowing us to remove all the image fetching code from
  atc/worker/
* resource type images can be fetched in the same way, removing the
  awkward state where a resource type has no version detected, and
  removing the need for resource types to have their own check builds.
* generally speaking, there's a lot of potential here for allowing parts
  of the build plan to be dynamically generated at runtime.

note: it is the responsibility of the caller to send an event containing
the plan so that it may be rendered in the UI, since the plan will not
be in the build's original plan.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>