vars: structure: derive Reference name
...rather than tracking the raw name that's passed into the template. It
felt a bit odd to me that we tracked this name, given that we never
really used it (practically only for error messages). It was also a bit
worrisome that we only sometimes set the Name, which felt like a
potential bug waiting to happen if we ever decided to use the Name for
something.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>