vars: structure: allow converting KV pairs to/from static vars
concourse/concourse#6041

Adds a new KVPairs type that is a list of key-value pairs. KVPairs can
be Expanded into a StaticVariables (which supports the usual Get/List
methods of a Variables impl), and StaticVariables can be Flattened into
a KVPairs.

The intention of Expand is to go from a list of var flags to a
StaticVariables for interpolating the pipeline template. The intention
of Flatten is to convert a pipeline's instance vars to a flat string
form using dot notation.

Note that we already have an implementation of both methods in
atc/pipeline.go for instance vars - however, I intend to use this
simplified implementation that does not traverse lists since the
notation for doing so is ambiguous (if a field name is an integer, it
assumes that it's an index in a list). Instead of using dot notation for
lists, we'll just display the list as '[a, b, c]'

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>