fly: structure: add InputMappingPairFlag
concourse/concourse#6041

Previously, fly execute was using the VariablePairFlag to specify input
mappings. Since we simply parsed key-value pairs as strings, this worked
fine - but, as I intend to make the VariablePairFlag smarter about the
keys (change them to vars.Reference), we'll need a separate flag for
input mappings.

I also added a helper to parse key-value pairs, since we have a whole
bunch of flags that do similar things.

This is technically a small behaviour change as it changes some of the
error messages.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>