atc: structure: add FetchImage to step delegate
I thought long and hard about this, 'cause these delegates are kind of
turning into god objects. But in the end I drank the kool-aid. Having it
on the delegate is super convenient. These delegates may be a bit
conceptually weak, but they at least make the steps a little easier to
read (and test).

...but we should definitely keep trying to think of something that makes
more sense.

...and there should *probably* not be one delegate per step type, since
then the boundary between them is even more fuzzy.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>