web: behavior: tuck away image check/get
adds an 'initialization' toggle button to the step bar, all the way to
the right, just before the status icon. placing it there kind of makes
sense because the status icon tooltip actually shows shows how long
initialization took, too. maybe a future iteration could split the
durations between the buttons.

when the button is clicked, the initialization steps are revealed in a
separate 'step body' with a slightly darker background (hooray 50 shades
of grey) so it appears distinct from the step body.

i re-used the 'cogs' icon from the build preparation dialog. i
rationalized this arguably lazy misuse by thinking of it as step
preparation. it might be a bit bright.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>