web: structure: optimize processing watch events
by only recomputing the pipeline layers when the inputs change. On a
large deployment, I found that the massive amount of updates to job
builds (e.g. builds starting and completing) were resulting in a ton of
CPU usage client side. There were ~100 updates/second. I speculate the
slowness was due to the pipeline layers computation, which would run
for each pipeline containing a job that got updated.

With this optimization, job updates that don't affect the job's inputs
do not trigger a recomputation. I suspect this'll be the most common
case - the shape of a pipeline isn't likely to change as frequently as
the jobs run

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>