web: behaviour: close event stream on errors
I deployed an image to the load test environment with a bug (the wrong
JSON was being emitted to the event stream), so Elm was erroring and
hitting the "fallback to polling" path. Unfortunately, we never actually
closed the event stream. This meant we were hitting the list all jobs
endpoint every time an event batch came in (which was 5 times a
second!), all while keeping the event stream open. This makes the UI
more resilient in that case

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>