atc: behavior: respect resource type version
the long term goal is to eliminate VersionedResourceTypes along with the
periodic resource type check builds, but while we're still in this
state, let's at least respect the version.

side note: later on when VersionedResourceTypes is removed we'll want to
find a way to ensure build re-runs use the same versions for these
images, but this problem already generalizes to task image_resources
too.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>