fly: behaviour: fix URL in set pipeline helper text
Since we modified the instanced pipeline query params in #6105, the fly
helper text should change too. This commit adds a new WebQueryParams
method to PipelineRef, which parallels PipelineRef.QueryParams for the
UI (rather than for the API). We may wish to converge to a single query
parameter pattern, but for now, lets keep it separate.

Also, we're now ignoring `found` when fetching the pipeline - we created
it immediately before fetching, so it should exist in practice. It
*could* have been deleted in the time between, but that's unlikely.

Signed-off-by: Zoe Tian <ztian@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>