atc: behavior: only send *public* plan to events
previously this would send the full private plan in all its sensitive
glory, storing it in the database forever. without encryption, this
could lead to storing sensitive credentials and even sending them over
the API in the build event stream.

this simply changes the events to contain the public plan instead. the
web UI still works just fine, since the public plans are a subset of the
private plans, and the web UI didn't use any of the old fields.

along the way, this also slightly tweaks some of the public plan `json:`
tags, since the various omitempty settings were wrong, and the order was
a little awkward.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>