atc: behavior: fetch task image using plan tags
this broke a topgun test which configured all workers to have tags. if
all workers have tags, and a plan does not specify any tags, it won't be
able to run on any workers.

the previous behavior (in <= v6.6.0) was to find a worker supporting all
the tags and the base resource type and fetch the image directly onto
the worker - now we'll just select a worker matching the same tags.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>