Fix k8s-topgun/worker_lifecycle_test
This commit fixes three things.

1. image checking now results in a get container appearing. When fly
hijack was being called it now got stuck waiting for user input. To get
around this we added the `-s` flag to pre-select the step we want

2. Correction when counting the number of workers. The test was checking
for workers not running, which it already does in the first Eventually()
at the top of the test. Changed it to check that there are zero workers
registered.

3. To produce less log output, we changed the task-waiting.yml pipeline
to sleep 10s instead of 1s

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Sameer Vohra <svohra@pivotal.io>