Fix topgun/k8s/baggageclaim real btrfs disk test
This test assumed that a persistent disk with a specific name already
existed in the GKE cluster and was ready to be mounted to any pod.
Someone appears to have cleaned up this disk though.

Refactored to instead do the following:
1. Create a storage class that creates btrfs disks
2. Create a PVC requesting a btrfs disk is made by the previous storage
class
3. use the PVC in the concourse deployment
4. Disk will be cleaned up because the ReclaimPolicy on the SC is
DELETE by default

Now this test should pass on any random GKE cluster that we run
topgun/k8s on.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Sameer Vohra <svohra@pivotal.io>