web: behaviour: switch back to old routes
concourse/concourse#5921

At the start of this PR, I had some large commits that switched the
routes from /teams/:team_name/pipelines/:pipeline_name to
/pipelines/:pipeline_id (sim. for the API endpoints), as it's more
convenient to work with just the ID. We ended up not wanting to do that,
because we don't want to get rid of the "friendly" routes.

So, here's a massive commit that switches everything over to the old
routes. The raw instance var JSON is specified in the query string of
the routes for pipeline instances.

Apologies for the massive commit - it's not so easy to split it up into
more managable chunks.

This reverts commit 753f8d3ebbdce9187be4ba1fea415041e28a2c0f.
This reverts commit a24bfee1b71285cb0357c4083123b51f9e8fbc54.
This reverts commit e8cb2cd1d4a2d49ed0022c9fa5e32b7575a44d92.
This reverts commit 4c3644fe5b10e59c64d5b0d4af51e9cfc551665f.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>