atc: behavior: update last check end time on fail
previously the check step would only update the last check end time when
the check succeeded. this meant that if the check failed or errored,
we'd actually check again immediately, which is a bit too aggressive.

this brings it back to the old behavior, which is to update the last
check end time on failure/error.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>