atc: structure: use vars.String for webhook token
Side note - it'd be nice to get rid of all of the creds.NewFoo helpers,
which would require changing each of the Foo types to an appropriate
vars.* type. One problem with that is that we'd lose appropriate
validation of structure for when the type is statically defined, unless
we made a custom vars.* type for each type that could be interpolated

e.g. if we made Source a vars.Any, a user could put a number in the
resource config YAML, and only complain at step runtime rather than
during set-pipeline

I have some ideas for this, but it gets a bit tricky with custom
UnmarshalJSON implementations

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>