web: fix non-trigger edges showing up as trigger
i set omitempty for the trigger field, which resulted in `=== false` not
being satisfied, since it became `undefined`. so let's just switch it to
`!== true`.

for anyone wondering why this isn't a simple bool check: the CSS sets
`stroke-dasharray` for non-trigger inputs, which could be set as the
default, but it's not super obvious what you'd set to remove the
dasharray when the .trigger class is present.

i guess you could set `.edge:not(.trigger)` to have a dasharray, but
don't really have an appetite to rock the boat in "legacy" CSS/SVG
logic.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>