atc: behavior: skip events bigint cleanup for now
also: during down migration, update newly written bigint-only events.
just do this in-place, assuming it won't be *too* many rows.

it's not really safe to run this because it makes it impossible to run
the 'down' migration. let's just leave this extra data around for now
and address it in a future release. it kind of seems like this could be
turned into a feature integrated into the migration library so it's less
awkward to handle, but that's a voyage for another day. (we should
probably also integrate voyager btw.)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>