web: behaviour: use proper types for pipeline jobs and resources
So this may seem like a structural change where we get rid of some
cruft, and it mostly is - just parsing the API responses as
Json.Encode.Value (as we did before) does no validation on the response,
and given that we actually do decode some of the values later, it's very
awkward.

However, I believe this fixes a very subtle bug that may have been
introduced by #6137. Occasionally, on the pipeline page, I'd notice that
the pipeline graph would stop updating (a job that I knew was finished
appeared to still be in flight). I looked at the dev console, and saw:

Uncaught TypeError: Cannot read property 'id' of undefined
    at K (elm.min.js?id=796bac482c493d401ada3762ee72d74f:formatted:208)
    at K (elm.min.js?id=796bac482c493d401ada3762ee72d74f:formatted:208)
    ... (the same error over and over)

Pretty cryptic. I looked at the `K` function that was erroring, and I
believe it was the Elm compiler's equality function (though it was also
cryptic as it was minified).

Googling Elm runtime errors, I found https://medium.com/@eeue56/top-6-ways-to-make-your-elm-app-crash-at-runtime-562b2fa92d70

"My rule: never do ==, toString, etc on Json.Value" (as it can lead to a
crash at runtime)

So, this is just a shot in the dark, but it does seem fishy that we're
getting an Elm runtime error at all. I suspect it has something to do
with the fact that resources now have a build field that can be `null`,
as there have been some runtime errors related to Json.Encode.null in
the past according to https://github.com/elm/core/issues/377. I haven't
figured out a repro yet, but even if this isn't the issue, it can't hurt
to get rid of some cruft!

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>