atc: behavior: timeout doesnt include image fetch
including image fetching in the timeout is odd because it would give the
task significantly less time to run if the image had to be fetched,
while also requiring you to allow more time for the image fetching
itself.

so rather than wrapping task/get/put steps in a timeout step, add it as
a field of each of them so they can enforce the timeout more narrowly.

for all other steps, the timeout step will wrap it and be enforced as
normal.

nb: this may be a "technically correct, but maybe not as useful" kind of
situation.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Co-authored-by: Bohan Chen <bochen@pivotal.io>