atc: structure: enforce timeout in check step
rather than having RunCheckStep take a timeout, enforce the timeout on
the ctx given to it like all the other steps.

note that this removes the custom timeout error message, but it was
inaccurate anyway, since the timeout may have been enforced externally
with a different value than the given check timeout.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Co-authored-by: Bohan Chen <bochen@pivotal.io>