atc: behavior: core step timeouts fail, not error
this matches the behavior of the existing timeout step. a testflight
test caught this! (and the check timeout test had to be updated so that
it's consistent with the other timeout tests.)

the errored state is reserved for unexpected errors. if a timeout is
configured, the timeout errors are expected, so it's considered a
failure.

side note: it's a little weird that this works by sending an Errored
event, but that's the way things worked before, and this commit just
aims to unbreak it. it might be better to emit a Finished event with the
actual exit status of the process, but we'd have to decide whether the
step should fail even if the exit status is 0.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Co-authored-by: Bohan Chen <bochen@pivotal.io>