web: behaviour: quoting filter does exact match
concourse/concourse#3010

When playing around with it, I noticed that it was a bit awkward to
force the term to be quoted on both sides - when you're typing a query,
you'll initially just have the opening quote. In this case, we still
want to filter, but the filter becomes a prefix filter.

There's also some awkwardness around having a single quote, which occurs
right when you start typing - without the extra check, we end up hiding
everything because the search term ends with a quote.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>