web: behaviour: pipeline status ignore paused jobs
If a job has been paused for some reason (e.g. it's known that the tests
are currently failing), the status of the paused job shouldn't
contribute to the status of the pipeline. It's very discouraging and
confusing to see a pipeline where all the jobs are either green or blue,
but the pipeline itself is red.

Not really happy with the duplication of `givenTwoJobs(WithModifiers)`
function in the tests, but I'm not too keen on adding even more
arguments to a pretty large function. And since this function is only
used in the status-hierarchy tests, I'm inclined to just leave it there.

Signed-off-by: Bohan Chen <bochen@pivotal.io>