web: behaviour: ignore leading white spaces
if a search query began with a space character then filtering would
fail. The error was from `symbol "-"`.

The `oneOf` would start down the first of the two branches, at which
point it was too late to back out and would error when trying to parse
for a dash "-" when there was no dash.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>