fly: optimize version checks test suite
The goal of the Versions Checks suite is to validate the behaviour for
version discrepency between fly and the atc. Previously, this was being
tested by keeping the atc mock servers's version constant, and changing the
version of fly in each test. Since fly isn't mocked out it has to be
rebuilt each time which even a single repetition of the build process
considerably (5x by our count) slowing down the whole suite. This slow
down can cause jobs on CI to timeout.

The test coverage can be achieved by keeping the fly version constant
and changing the atc mock server's version each time. This means fly
will only have to be built once speeding up the suite significantly.

concourse#6291

Signed-off-by: Rui Yang <ryang@pivotal.io>
Co-authored-by: Rui Yang <ryang@pivotal.io>