atc: behavior: simplify sql migration handling
rather than parsing out all the statements and supporting special syntax
for transactionless migrations, just evaluate the whole content.

the parsing was super brittle and would break if you used comments or
special syntax in certain ways.

we've been consistently adding BEGIN ... COMMIT in all our migrations,
so it doesn't seem worth the hassle; if you need it to not run in a
transaction, just don't add the BEGIN / COMMIT.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Co-authored-by: Muntasir Chowdhury <mchowdhury@pivotal.io>