containerd: behaviour: implement container.Info
Currently, it returns only the IP address of the container's virtual
ethernet pair, as well as any additional properties that are set as a
raw map (e.g. grace time) - so it's not Garden compliant, but Concourse
clearly doesn't need it to be (we currently don't even use the container
info endpoint).

The motivation for this can be seen in the newly added integration test:
finding a containers IP address so that containers on the same host can
communicate via the bridge network. This will make concourse/rfcs#84
possible with a containerd runtime.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>