atc: add first pass at get var step
This is a wip commit with the first pass at the get var step. It is
implemented using the existing credential fetching logic. We added a
global cache to cache credentials that are fetched through the get var
step. The result of the get var step is adding the var value to the
build vars, which then can be used by other steps.

The build vars was refactored to be able to include vars that are not
local (meaning you can save build vars within sources other than ".").
This is so that the values of the get var step can be used within other
steps without referencing the local vars.

[#6274]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>