rename migrations to accomodate v6.7.2 release
v6.7.2 introduced a migration whose timestamp was newer than these, so
they need to be bumped so they'll actually run when someone upgrade to
v7.0.0.

we've already run all these on ci.concourse-ci.org, so we need their
timestamp to be newer than v6.7.2 but older than the latest version that
ran there.

closes #6277

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>
Co-authored-by: Bohan Chen <bochen@pivotal.io>