generate example pipelines from github
when we change the example pipelines, it's easy to forget (or just not
be bothered) to update the example in the docs.

this commit adds a new `remote-codeblock` command that generates a
codeblock by making an HTTP request to the provided endpoint. this lets
us point at the examples repo directly so the pipeline configs stay in
sync

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>