atc/worker: fix limit-active-tasks
Limit active tasks is supposed to wait for a worker to become available.
This broke in #6339. This commit fixes it when limit-active-tasks is one
of the placement strategies.

Also added a test to ensure the waiting for worker path is executed.

I tried to modify the code so all strategies would wait for workers but
I think this kind of waiting needs to happen in the worker.pool which
would break the metrics for limit-active-tasks which happens in
worker.client.chooseTaskWorker().

Signed-off-by: Taylor Silva <tsilva@pivotal.io>