atc: behaviour: select worker in step execution
rather than in the worker.Client. the worker.Pool now gives a
worker.Client from the SelectWorker method, which makes it possible to
support different runtimes in parallel in the future (though, the
worker.Provider will need to handle this and return the appropriate
worker.Client instead of a worker.Worker).

selecting a worker for task steps has been moved to the TaskDelegate
since there's some funky logic around limit-active-tasks. I also
noticed/fixed a bug in this logic wherein the limit-active-tasks DB lock
would not be released under some failure cases.

side note, but this code really feels like DI/mock hell - it'd be nice
to clean this up/reduce the indirection, but I'll leave that for another
PR.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>