atc: behaviour: deterministic ordering of ?vars params
when parsing the key-value pairs for pipeline instances from API
requests.

this is unlikely to matter, since internally, we only use the query
params in two ways:

* ?vars={...full JSON...}
* ?vars.foo=123&vars.bar.baz="def" (i.e. every var in non-overlapping)

initially I figured we could just use the order that the query params
were specified in, but since `url.Values` is a `map`, iteration order is
non-deterministic. while we *could* use `url.RawQuery` and parse it
ourselves to maintain the order, this didn't seem worthwhile

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>