atc: refactor first occurrence query
This should help improve the speed of this query to find if a version is
the first occurrence within a job. Previously, the query would fetch all
the builds within a job and then try to find a version within those list
of builds that match the version. But that causes a lot of slowness when
the job has a looot of builds. (the CTE part of the query is slow)

The fix here is to have the first part of the query (the CTE) to find
all the builds that have that version and then try to find a build that
matches the job we are looking for. This will hopefully help reduce the
number of rows from the CTE, because there are probably less builds that
match the version than there are builds that are from the job. One
concern is that there can be some versions of a resource that are being
used by many builds within a pipeline, which can cause the query to be
slow too because it needs to fetch all of those rows. But at least in
the most common case, I believe that there will be less rows from
narrowing down with a version vs narrowing down with a job id.

Signed-off-by: Clara Fu <cfu@pivotal.io>