web: structure: use BuildStep to render step header
...rather than the StepHeaderType, which carries no additional
information.  doing this means we can access data for the step beyond
just the name to display in the header. the across step already benefits
from this slightly - rather than generating a pseudo-name, we have
access to the `vars` in the `AcrossPlan` and can generate the header
from that - e.g. if we chose to wrap each var in a `span`, we could do
that now

more importantly, though, we can display more structured info in the
set_pipeline step header - while we could try to stuff the instance vars
into a single string, that's less ideal since we can't apply styles to
them

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>