web: behaviour: fix pipeline lookup for instance vars
when we archived the prs pipeline and added some prs instance pipelines
(with the same name as the old pipeline), the UI was rendering the pages
as if the pipelines were archived. this is due to a bug in the pipeline
lookup logic to check if the pipeline is archived - we were only
considering the team name and pipeline name for uniqueness

since we now have a common function for looking up pipelines
(SideBar.lookupPipeline), let's use that instead.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>