Add a flag to migrate to the latest db version
This commit adds a `--migrate-to-latest-version` flag to the `migrate`
command. It will migrate the db to the latest version.

While doing concourse/concourse-chart#194 we found that the migrate
commnd does not automatically migrate to the latest db version. You can
do some shell scripting to accomplish that task using migrate and its
flags. We didn't want to write a script to do this in the chart though
and this seemed like a useful thing to add to the migrate command.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>