atc: wip plugging in get var as a delegate method
We want to have all the evaluate of credentials to use the get var step
rather than what it is doing currently by evaluating in memory within
the atc. In order to do this, we want a sub step to spawn whenever a
step needs to evaluate any credentials. So when a step calls Evaluate,
this will result in a get var step being run as a substep.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>