atc/db: Use ContainerRegistry to get number of active containers
There is a differential between the number of containers reported by fly
workers and fly containers. The workers table tracks the number of
active containers separately from the containers table which feels like
a duplication of effort. By getting the consumers to check the
containers table directly, we can get an equal reflection of how many
containers are running, especially starting containers as the ATC
updates the containers table directly rather than waiting for a
heartbeat from the worker to report the new containers. If there is a
lag in updates due to the container sweep only being every 30s, a
separated "report containers" sweeper could be setup that runs more
frequently and remove the synchronous "report containers" run from the
main container garbage collection sweeper.

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>