web: behaviour: quoting filter does exact match
note: this is pretty much exactly copied from concourse/concourse#6245

this adds two new string filters - Exact and StartsWith. Exact is when
you have a string enclosed in quotes (i.e. "..."), and StartsWith is
when you just have the opening quote, i.e. "...

StartsWith is intended to help you while you type a query by filtering
suggestions by only those that could match exactly if/when you decide to
close the quotes for an exact match

this also quotes the suggestions for teams, meaning if you select a team
suggestion, it'll only show that team's pipelines - which is likely what
you want

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>