web: behaviour: add group filter
for filtering by an instance group name. this currently behaves exactly
as the pipeline name search, except it filters out non-instance groups.
it will later be extended such that the dashboard view changes to the
"instance group" view when this filter is applied. this extension will
allow viewing an instance group to be a special case of filtering. it
also will mean you can view multiple instance groups at once.

this commit removes some duplication in computing the instance groups,
and changes the dropdown suggestion tests to not call Filter.suggestion
now that we just pass in the filtered pipelines to suggestion (so
testing Filter.suggestion alone isn't all that useful)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>