atc: fix exec tests and integrate creds
Originally the get var step was just creating a new instance of each
credential manager and running the get off of it to grab the value of
the var but this wont be enough if we want to ship this as a real step.
This commit starts on integrating the logic that we already have around
using credential managers like having a pool of cred managers or the
existing manager caching that should be removed for var sources. ALso
integrated the get var within the delegate to look at the global cred
manager if the source is empty.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>