Use X-* headers for getting a client's IP address
Getting a client's IP address is a bit harder than initially thought.
There are two headers that one should check before falling back to the
RemoteAddr available in the request (which is guaranteed to always be
set btw).

The auditor will now check for X-REAL-IP and X-FORWARDED-FOR headers and
log those, in that order, as the client's IP. This is not foolproof and
is easily spoofed by a client.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>