atc/db: Get worker.ActiveContainers from containers table
Rather than tracking the active container count via the heartbeat, just
use the containers table instead. This gives a more accurate
representation of how many containers there are from the ATC's point of
view since this is what is used for scheduling decisions and brings the
`fly workers` and `fly containers` numbers in sync. This data is
refreshed from the workers slightly less frequently than the heartbeats
but is updated by the ATC as soon as a container is created so has
faster initial feedback which may help with scheduling decisions.

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>