atc: add var source configs to run state
This commit tackles the previous problem of needing to propagate a
modified list of the var source configs to a possible get var substep.
This is because when we are evaluating the source of a var source within
a get var step, that var source's source might also need to be evaluated
which will result in running a get var substep for it. The list of var
sources that get passed to the get var substep should not include the
var source that we are trying to evaluate. In order to do this, we are
passing in a list of var source configs into the Variables method on the
delegate. This list of var source configs will then be used to update
the var source configs saved on the run state of the child scope that
will be created for the get var sub step.

This way of updating the var source configs through the Variables method
is prooobably not the best? It is the only thing I can think of right
now but there might be a better way to do this.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>