atc: add back local variables and implement caching
This commit changes the build variables to only contain local variables
set by the load var step. We previously changed it to hold all variables
that were fetched by the get var step, regardless of if it was fetched
by the global credential manager, load var step or a var source. But we
realized that build variables are persisted throughout the entire run of
the build, even if they are expired. I originally thought that this was
what we wanted but actually we should only be using a var value if it is
not expired. So that means we would need to implement an expiration to
build vars, which then removes the purpose of putting it on the build
vars in the first place because we already have caching that does the
expiration logic for us. We decided to just store local vars on the
build variables and rely on caching to deal with storing global and var
source fetched vars within the get var step.

One thing to note is that vars fetched from global cred manager will be
using the old way of caching (not through the get var step, meaning
won't be stored in the global cache), while any vars fetched from get
var step (var sources) will be stored in global cache. The global cache
will be used for all vars but the team id key will be added to the cache
key to avoid teams from sharing credentials.

Because we are no longer adding vars to the build vars during the get
var step, we needed to add tracking to each var within the get var step
manually.

I also implemented caching properly using the vars returned from the ATC
configured fields. It will take into account whether caching was enabled
or use the default values that are set.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>