atc/vars: vars are tracked with source and path
Previously we were only tracking vars using their path but that could
result in a bug if there are multiple vars with differnt var sources but
the same path. This commit fixes this and has vars being tracked by
their source and path.

Also implemented tracking onto the run state so that vars can be tracked
using a Track method on the run state. The same tracker will be used to
track local vars. When a new scope is created, it will construct a new
tracker that still maintains a relationship with the parent's tracker.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>