atc/creds: add back registring manager factory on init
I realized that we actually still need the manager factories to be
constructed on init because we don't have a full list of all the manager
factories anywhere. The only thing we have is the full list of managers
within atccmd, but that doesn't give us the manager factories. We need
the list of manager factories for constructing the cred manager for var
sources.

[#6274]

Co-authored-by: Clara Fu <cfu@pivotal.io>
Signed-off-by: Clara Fu <fclara@vmware.com>