flag: add the flag package to the concourse repo
This flag package is very similar to the github.com/concourse/flags
repo but with some modifications and additions to have it support being
used as a Value interface from pflags. It needs to support UnmarshalYAML
for unmarshalling the config file into the type and also Set, String and
Type for pflags to decode the flag value into this custom type.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>