atc: fixes errors after merge and adds env values to fields
In order to keep the env var values backwards compatible after moving
around and renaming the fields in the flag structs, I added two options
to the construction of the env var name in the envstruct package. The
two options are `ignore_env` and `env`. If ignore_env=true is passed in
the tags for that field, envstruct will not include the yaml tag name in
the built up string to use to fetch the env. If env is passed in then it
will override the built up string to fetch the env using the yaml tags.
This will allow us to keep the old env names so that everything remains
backwards compatible, but also still allow the new env names to be
used.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>