B: generate static plans for image sub steps
[#6274]

This is a squashed commit that included a TON of commits so I'm sorry
for the terrible commit message!

Previously, we pushed the creation of image check and get plans into the
planner but did it through individual steps wrapped together through on
successes. This brought some complications such as needing to change how
it is rendered in the web UI and also made it difficult to read the build
plan. This commit keeps the creation of the image check and gets in the
planner but rather than running them separately and tieing the step and
it's image check/get through the ImageFrom field, we inline the image
plans directly in the step itself. Because of this, all FetchImage
within each step that needs an image only needs to run the
preconstructed plan and retrieve the results from the run state.

There will be an ImageGetPlan and ImageCheckPlan added to each step plan
that requires an image. Each step will run these constructed plans,
similar to how it works now but only difference is it no longer needs to
create the plans.

The logic for the current FetchImage needs to be moved into the task
step since it still needs to create substeps for dynamically fetched
task configs.

Image check steps store versions in the runstate, which is then grabbed
by the get step and used. We modified the version grabbed from the get
step to not be specific to the put step, since we are now using it for
check steps as well. The planID specified by ImageSpecFrom actually
stored more results than just the ImageSpec, so let's make the name more
accurate.

The Check step stored just the version, but the Put step stored the
version and the metadata. We couldn't find a use for the
`VersionResult.Metadata` anywhere so let's only store the Version
instead

- add test coverage, most of the image spec generation logic will be
tested in the get step tests
- it would be painful to maintain different codepaths for when the
task config is known vs when it's dynamically fetched

TODO:
- pass the varsourceconfig through the scheduler into the build plan
- figure out why we're storing runtime.VersionResult, do we actually use
the VersionResult.Metadata in any subsequent operations?
- Because we removed generating new scopes in FetchImage, it's now
possible for artifacts generated by get step to collide with image get
steps. The solution we're thinking of is to split the artifact
repositories into 3, resource, resourceType, and anonymousResource. This
way we can rely on the guarantees provided by ConfigValidate (i.e. no
duplicate resource or resourceType names) and flatten the run state
without worrying about if a get for a resource type named "foo" will
collide with a resource named "foo".

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>

[#6274]