add omitempty to all config fields
The omitempty will make it so that the field is not included when we
yaml marshal the config from the struct into a yaml string. I was hoping
that in the future we would have a command that you can run in order to
have concourse return back the yaml config file it is currently using to
run. Also it will help with the migration path for users going from
flags/env to using a config file.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>