flag: saving original value and handling nil values
For some flags such as private key and cipher, we use the field that
they pass in and convert it into the type that we need. The original
field is not saved and there is no way of converting the type back into
the original field. When we want to convert the concourse struct back
into a yaml file, we need the original fields that were passed in. In
order to have those, we need to save the original fields along with the
converted type.

Also we are handling cases when the value of these custom types are nil.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>