web: behaviour: reintroduce limit-active-tasks placement strategy
The "limit-active-tasks" placement strategy was temporarily removed during refactoring, and this commit adds it back.

NOTE: The "limit-active-tasks" strategy is a bit less precise now, similar to "limit-active-containers" and "limit-active-volumes", as there is a small period of time between checking the active task count for a worker and picking / incrementing the active task count for the worker (if chosen). Should multiple tasks attempt to be scheduled at the same time, they may all pick the same worker and push the active task count for that worker above the configured limit.

Signed-off-by: David Symons <david@symons.io>