S: rename ImageSpecFrom and store Version in state
The planID specified by ImageSpecFrom actually stored more results than
just the ImageSpec, so let's make the name more accurate.

The Check step stored just the version, but the Put step stored the
version and the metadata. We couldn't find a use for the
`VersionResult.Metadata` anywhere so let's only store the Version
instead

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>