B: redo FetchImage so it doesn't generate substeps
[#6274]

Instead, because of the changes to the scheduler, the build plan
should've created any dependent check/get steps and set the proper
references. All FetchImage needs to do now is to retrieve the results
from the run state and construct the image spec accordingly

The logic for the current FetchImage needs to be moved into the task
step since it still needs to create substeps for dynamically fetched
task configs.

TODO:
Because we removed generating new scopes in FetchImage, it's now
possible for artifacts generated by get step to collide with image get
steps. The solution we're thinking of is to split the artifact
repositories into 3, resource, resourceType, and anonymousResource.

This way we can rely on the guarantees provided by ConfigValidate (i.e.
no duplicate resource or resourceType names) and flatten the run state
without worrying about if a get for a resource type named "foo" will
collide with a resource named "foo".

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>