atc: inline the image get/check plans
Previously, we pushed the creation of image check and get plans into the
planner but did it through individual steps wrapped together through on
successes. This brought some complications such as needing to change how
it is rendered in the web UI and also made it difficult to read the build
plan. This commit keeps the creation of the image check and gets in the
planner but rather than running them separately and tieing the step and
it's image check/get through the ImageFrom field, we inline the image
plans directly in the step itself.

There will be an ImageGetPlan and ImageCheckPlan added to each step plan
that requires an image. Each step will run these constructed plans,
similar to how it works now but only difference is it no longer needs to
create the plans.

[#6274]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>