B: implement resource getter for running get steps
this re-implements worker.Fetcher and worker.FetchSource for running get
steps (under a lock) if necessary (i.e. the resource is not cached on
the worker). this should not change the behaviour of Fetcher/FetchSource

introducing bogus interfaces for the sake of easier testing is something
I tried to avoid in this refactor, but the exec tests are already a bit
of a mess and may require a large refactor of their own - until then,
I'm gonna test against this interface in get_step_test to keep my sanity

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>