add locking to new build variables file
In the get var epic, I moved the build variables to a new file but in
master there were changes made to add locking to the access of the
variables map. This commit adds those changes to the new file. Also
fixed some compiler errors from the previous changes made to rename
BaseImageType to BaseType.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>