B: allow configuring the password connector
currently the only supported connectors are local and ldap. the username
prompt for ldap can be configured - this appears on the login with
username/password screen (e.g. setting it to "Email" will replace
"Username" with "Email"). we had to hack around Dex's default username
prompt for the "local" connector, since it defaulted to "Email Address"

this commit also changes the style of the login with username/password
screen

Signed-off-by: Rui Yang <ryang@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>