move validator package
I moved the validator package from atccmd to the cmd package because I
want to have it be able to be used for validating the fields from
packages outside of the atc. I needed to add validation for fields
within the worker/baggageclaim packages such as log level, ip
versions and baggageclaim driver.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>