simple checking if manager/emitter is configured
Previously, we were checking if a credential manager or metrics emitter
is configured by using the reflect package to grab each field off of a
struct and converting each field to an interface to see if it is
configured. I've been debating whether this was the best way to do it or
not because it does add a layer of magic between the managers/emitters
and where/how they are being used. The good thing about this way is that
people adding new managers/emitters will not need to know where to plug
their new addition in, they will just need to add it to the appropirate
struct. But I decided that less magic is worth the risk of someone
forgetting to add it to certain places, and I added a test that will
loop of each field in the struct to make sure people can't forget to add
it to the method that checks if the manager/emitter is configured.

Also in this commit I moved the config types into a separate file just
to keep it cleaner.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>