web: behaviour: Proactively wake step waiting for worker
Attempts to proactively wake a step waiting for a worker to become available. It isn't guaranteed that the step which is woken will be able to run on the released worker, but if it can this can cut down the amount of time wasted between releasing a worker and the next tick where steps check for available workers.

Signed-off-by: David Symons <david@symons.io>