convert skymarshal to use new config tags
Converts all the skymarshal connectors and team connectors to use the
new tags for the config file parsing of web fields. It also moves the
validation of the connectors within the DisplayIDPerConnector field into
the validation package rather than a runtime method within the skycmd.

One thing to note about this is that the set-team command in fly still
requires the old go-flags tags for each of the team connectors because
fly is still using go-flags. I think we will be converting fly to use
cobra in the next step of this epic so for now, there will be a lot of
old go-flags tags within the skymarshal team connectors fields until fly
is converted to use cobra.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>