B: fix font antialiasing
minor thing, but due to changes in our dependencies (probably the less
bump in #6698), fonts started appearing thicker than usual.

this appeared to be due to an improper value for -webkit-font-smoothing.
in our `.less` files, we quoted the value, which is invalid. I suspect
older versions of less were smart enough to strip the quotes, but the
new version no longer does that.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>