add configured connectors and fix atc tests
I realized that the DisplayName field on multiple skycmd connectors were
not only used for the help text, but also the name that shows up in dex.
So I added it back and then properly used the new way of grabbing
connectors (directly through the atc config rather than the global list
of connectors) to set the connectors on the dexserver. Also added a test
to the connectors config struct in order to make sure we are checking
each connector is checked in the ConfiguredConnectors method.

This commit also fixes the remaining tests in the atc package.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>