add tsa command and separate flag files
Added a new tsa command in order to run the tsa independently from the
web concourse command. Also separated out the atc and tsa deprecated
flags into its own independent files within their respective packages.
This is so that we can have the tsa command call the same method in
order to set up the cobra flags.

Also fixed up the tsa cmd tests to get them working with the new config
file changes. One thing to note is that I guess starting up the tsa
using the new method of a config file with cobra is slower than using
the go-flags package so I had to up the StartCheckTimeout to 1 minute.
Not sure if there's anything that can be improved there to make it
faster so we don't have to set a higher timeout. When running the
tsa/cmd tests on my local machine, it used to take around 200 seconds
and now it takes 230 seconds.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>