S: Drag/DropState keeps track of Card
...instead of cardIdentifier and teamName.

previously, we were using cardIdentifier to keep track of what card we
were dragging/dropping over. this makes it harder to access information
about the cards - for instance, when dragging an InstancedPipelineCard,
we don't have an easy way of figuring out what instance group the card
belongs to - we'd have to either add the pipeline name on DragState
redundantly, or lookup the card by its identifier.

storing the full card that we are dragging makes this easier, since no
lookup is required. it also lets us remove the teamName on the
DragState, since that information can be inferred from the card

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>