B: wrap stdin when using containerd with TTY
[#6632]

containerd treats closing of stdin as a SIGHUP to the container. But
because we're forwarding stdio over the internet, we might run into
connectivity issues.

the current solution is to assume 1. we only use stdin at the very
beginning of a container (resource scripts), and 2. the network flake
will probably occur after we've already sent everything we needed.

the simplest solution is to just wrap the stdin in a reader that never
returns (until the process exits).

Signed-off-by: Bohan Chen <bochen@pivotal.io>