add log level flags and proper usage of cmd defaults
Added forgotten log level flags to the atc and worker commands. Also
converted the worker flags to use the dedicated defaults specified in
the worker CmdDefaults. The CmdDefaults should be the source of truth
for anything that needs to reference a default value.

Also made it so that baggageclaim has its own validator and so I moved
over all the validations that are on behalf of baggageclaim to the
baggageclaim package and we will reference them from there.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>