create check builds as started
...rather than creating them as pending first and then starting them. we
assume this was done to avoid duplicate logic when starting a build, but
this seems to have some performance impacts.

the `build.start()` UPDATE query consumes the majority of the CPU usage
on CI and hush-house. most of that CPU time is in the index scan on
pending_builds. since this index first indexes by job_id, but check
builds don't have a job_id, this ends up taking a while occasionally.

however, if we just create the build in the started state to begin with,
we don't need to do this index scan in the first place.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>