add exception handling in bigint migration
there is a race when bigint migration running the
pipeline_build_events_* table got deleted by GC while the loop
still tries to manipulate indexes of the table. The exception
block will catch the error if table is already gone and won't
try to drop/rename indexes if not exists.

Co-authored-by: Taylor Silva <tsilva@pivotal.io>
Signed-off-by: Rui Yang <ruiya@vmware.com>