add enabled field on credential managers
The enabled field will allow Concourse to know if a credential manager
is configured or not. Before we would check if certain fields are set in
order to know if the manager is configured, but a really common practice
in config files is having an explicit enabled boolean which will
determine if a field is set. One other reason is that within a config
file, it would be weird if only certain fields being set will determine
if a manager is set or not. The validate methods on each manager was
adjusted to include validations on the fields that were required to
check if a manager was configured or not.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>