add enabled field to auth connectors
The enabled field will allow Concourse to know if a auth connector
is configured or not. Before we would pass try to validate all the
connectors and pass the appropriate ones to dex but a really common
practice in config files is having an explicit enabled boolean which
will determine if a field is set. With this change, we will only pass
the enabled true connectors to dex, this is mostly to benefit the
consistency of how integrations with concourse are handled (make it the
same as metric emiiters, cred managers, etc).

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>