add enabled field on metrics emitters
The enabled field will allow Concourse to know if a metrics emitter
is configured or not. Before we would check if certain fields are set in
order to know if the emitter is configured, but a really common practice
in config files is having an explicit enabled boolean which will
determine if a field is set. One other reason is that within a config
file, it would be weird if only certain fields being set will determine
if an emitter is set or not. The validate methods on each emitter was
adjusted to include validations on the fields that were required to
check if an emitter was configured or not.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>