fix and chase errors
There were a few bugs here and there in the code that I fixed in this
commit.

One addition is that the config file is able to be set through env vars.
I did this because the docker-compose relies on env vars to run the
concourse command. Not sure how to make it run the flag, will prob leave
that to later because we do want to deprecate env.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>