remove enabled field from creds managers
The way that you configure a runtime for a worker is by setting a
runtime field to a specific runtime that you want to set and then
concourse will use the fields that are configured for that specific
runtime. This is a different method than choosing the option through an
enabled field. In order to keep all the option choosing methods the same
throughout the config field, I decided to switch over all the places
where the user has to choose an option to having them fill in a field in
order to determine which option they want to use rather than setting
enabled=true for the option. In this ex, they would set

credential_manager: vault
vault:
  field1: ..

Rather than previously

vault:
  enabled: true
  field1: ..

Next I will need to switch over the metrics emitters and tracing providers.
The auth providers do not need to be switched over because I will be
keeping the enabled boolean for them. This is because you can configure
multiple auth providers.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>