remove enabled field from tracing providers
The way that you configure a runtime for a worker is by setting a
runtime field to a specific runtime that you want to set and then
concourse will use the fields that are configured for that specific
runtime. This is a different method than choosing the option through an
enabled field. In order to keep all the option choosing methods the same
throughout the config field, I decided to switch over all the places
where the user has to choose an option to having them fill in a field in
order to determine which option they want to use rather than setting
enabled=true for the option. In this ex, they would set

tracing:
  provider: jaeger
  jaeger:
    field1: ..

Rather than previously

tracing:
  jaeger:
    enabled: true
    field1: ..

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>