error when failing on validation of auth provider
Before we would not exit with an error when a validation failed on a
configured auth provider because the validation is how we figured out if
an auth provider was configured or not. Now that we have an enabled
boolean to tell us if an auth provider is configured, we can use the
Validate method in order to throw an error whenever a configured auth
provider is configured incorrectly.

I also filled in the code for converting the flag values to the current
config format of all the integrations with concourse around auth
providers, metric emitters, credential managers and tracing providers.
This is because previously we would have IsConfigured methods on all
these integrations and within the IsConfigured method, it would check
if certain fields are set in order to determine if an integration is
configured or not. Now that we have a config file, we can have the user
write in which integration they want to use. For ex, previously the user
would need to set

--bitbucket-cloud-client-id=<client-id>
--bitbucket-cloud-client-id=<client-secret>

in order to have bitbucket be configured as an auth provider. Now we
want users to explicitly write which integration they want configured in
the config file

auth:
  provider: bitbucket_cloud
  bitbucket_cloud:
    client_id: <client-id>
    client_secret: <client-secret>

If we want to keep this backwards compatible, we need some way to
convert the old way of checking if an integration is configured into the
new way of having it be set in "provider" field. So this commit achieves
that by checking the exact fields that we used before to figure out if
an integration is configured, and if it was configured we will manually
set the "provider" field on the struct. If there are more than one
integrations set, we will error (except for auth providers where there
can be multiple).

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>