add inline to yaml tag for nested structs
I didn't know that marshaling and unmarshaling yaml will actually take
in fields that don't have the yaml tag and itll just try to find the
field name in the yaml file. I don't want that because we have many
nested structs that are only there for structural purposes in the code
and is not supposed to be represented in the yaml config file. In order
to keep those nested structs for structural code purposes only, I added
an "inline" value to the yaml tags to make the yaml
unmarshaler/marshaller ignore that struct field name.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>