fixup cmd files (esp quickstart)
The quickstart command was actually incorrectly built because I forgot
to add the initialization of all the flags to it. So this commit adds
all the web and worker flags to the quickstart command but because the
worker flags in quickstart need to be prefixed with "worker-", I had to
add a prefix to all the flag names when you call the method to
initialize all the worker flags.

Another thing that this commit does is change all the structs for ATC,
TSA and worker to non-pointer structs. I had it as pointer to the
structs initially because that is what we had done previously. But now
that we use envstruct, it cannot traverse the values in the struct if it
is an uninitialized pointer. So either I had to intialize the pointer or
just make it a non-pointer. I don't think there is a reason why we need
it to be a pointer so I just removed the pointer.

Last thing is that I removed the ability to fetch environment variables
for the TSA if you are running the TSA independently. This is because I
didnt previously notice that you were never able to use env vars to
start up the TSA independently, and I don't want to add that as a
feature.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>