remove descriptor from find or create resource config
Removed the resource config descriptor from the
FindOrCreateResourceConfig method on the resource config factory. We
were able to do that because we are now passing in the resource cache
for the custom type (if it uses one), and with the resource cache we no
longer require the descriptor to tell us what its chain of parent types
are.

Also updated a bunch of tests to respect the new resource cache
argument from FindOrCreateResourceConfig.

Currently FindCheckContainer is broken because it also used
FindOrCreateResourceConfig, but now that it takes in a resource cache we
dont have a resource cache to pass into it from the FindCheckContainer
method. We will be removing the need to get a resource config at all in
the future for FindCheckContainer so we can keep it broken for now.

[#6274]

Signed-off-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Clara Fu <fclara@vmware.com>