put step to use image resource cache for saving output
We have converted check and get steps to create a resource config or
resource cache using an image resource cache in order to avoid needing
to use the versioned resource types to create a resource config/resource
cache descriptor. This commit makes it so that the put step will
propagate the image resource cache to the db build in order for
SaveOutput to use to create a resource config. This way we can remove
VersionedResourceTypes from the put plan. Also filled in and fixed some
exec tests.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>