atc: remove check planner object
Originally, a resource or resource type db object would be used to
create a check plan. During this refactor, I changed it so that the
check plan would be created through a Check Planner within the builds
package. The reasoning was because I wanted all the logic for creating
build plans to be in one place because we were adding more complex
logic with building image plans. After discussing this with Alex, we
decided it might be better to move the creation of plans back to the
resource and resource type object. This is because the check plan that
needs to be created for a resource and resource type is slightly
different, and also it is better practice to keep a small interface for
a narrowly scoped method (for ex, creating an image plan) and to use
that small interface in any part of the codebase (rather than if the
interface was a check planner which holds a bigger scope).

With this change, we also refactored the added image fields
(ImageGetPlan, ImageCheckPlan and BaseType) into a struct labelled as
"image" in the plans. We also moved Privileged into the image struct
because the privileged boolean is related to the parent custom type.

The versioned resource types will now contain the method for
constructing an image plan. This is because it seemed to be the place
that makes the most sense for where it belongs, since the image plan is
for fetching the image of a custom resource type.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>