atc: remove check planner object
This is a squash of many commits so Im sorry if this commit is pretty
big :D

- Originally, a resource or resource type db object would be used to
create a check plan. During this refactor, I changed it so that the
check plan would be created through a Check Planner within the builds
package. The reasoning was because I wanted all the logic for creating
build plans to be in one place because we were adding more complex
logic with building image plans. After discussing this with Alex, we
decided it might be better to move the creation of plans back to the
resource and resource type object. This is because the check plan that
needs to be created for a resource and resource type is slightly
different, and also it is better practice to keep a small interface for
a narrowly scoped method (for ex, creating an image plan) and to use
that small interface in any part of the codebase (rather than if the
interface was a check planner which holds a bigger scope).

- We also refactored the added image fields
(ImageGetPlan, ImageCheckPlan and BaseType) into a struct labelled as
"image" in the plans. We also moved Privileged into the image struct
because the privileged boolean is related to the parent custom type.

- The versioned resource types will now contain the method for
constructing an image plan. This is because it seemed to be the place
that makes the most sense for where it belongs, since the image plan is
for fetching the image of a custom resource type.

- In the TryCreateCheck method, if the list of resource types is nil we
need to convert it to an empty list. This is because we are using the
versioned resource types as an object that conforms to the ImagePlanner
interface, and if it is nil then the code that calls a method off of
the ImagePlanner interface will panic.

- In the Image plan of a step plans (get, check, put and task), whenever
we need to fetch an image for a custom type we will now always create a
check step for the custom type. This won't have any performance
regressions because the check will only run if it has been less than the
checking interval, so we don't have to worry about a hoard of checks
being run for custom types. This will allow us to remove the need for
checking resource types explicitly in lidar.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>