atc: ensure versioned resource types is not nil
In the TryCreateCheck method, if the list of resource types is nil we
need to convert it to an empty list. This is because we are using the
versioned resource types as an object that conforms to the ImagePlanner
interface, and if it is nil then the code that calls a method off of
the ImagePlanner interface will panic.

Also plugged in the changes made to the structure of the image fields on
the get, check and put plans into the exec package.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>