atc: always create check plan for usage of custom type
In the Image plan of a step plans (get, check, put and task), whenever
we need to fetch an image for a custom type we will now always create a
check step for the custom type. This won't have any performance
regressions because the check will only run if it has been less than the
checking interval, so we don't have to worry about a hoard of checks
being run for custom types.

This will allow us to remove the need for checking resource types
explicitly in lidar.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>