switch map types to flag.StringToString or StringToInt
This has to be done because go-flags parses map type values with a colon
(ex. map-type=value:1) but with cobra that uses pflags it parses
map types using equals (ex. map-type=value=1). In order to make
it use colon instead of equals for backwards compatibility, we have to
make a custom type unmarshaler that will separate map key and value by a
colon. This commit replaces all usage of map[string]string and
map[string]int field types with the flag types of StringToString and
StringToInt custom types. Once we have fully deprecated flags we should
be able to switch them back to just the golang map types. Any new fields
added does NOT need to use the custom flag types and can use
map[string]string or map[string]int directly as the type.

Also converted field types that use a pointer to no pointer. This is
because it was not really necessary for them to be a pointer, and it
being a pointer value makes the flag parsing more complicated.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>