atc: remove atc.VersionedResourceTypes
Since we no longer use the version off of the VersionedResourceTypes
(because we always create a check whenever a custom type is used) inside
a build, we can remove this type entirely and replace all usages with
ResourceTypes. This commit replaces all the places where
VersionedResourceTypes was used within the ATC. We still need to convert
go-concourse and fly to no longer use it.

This might have some performance benefits because we no longer need to
query for a version for each resource type when we fetch the resource
types from the database.

Also removed versioned resource types from fly and go-concourse. A few
major changes within this commit:

- go-concourse: change VersionedResourceTypes to ResourceTypes. The
resource types returned from that method will not longer contain the
versions for each resource type, which should be okay since nothing used
the version.

- fly execute will construct a TypeImage within the GetPlans for each
input. This way fly execute will be able to construct the correct build
plan for each input.

- fly check-resource and check-resource-type no longer run an individual fly
check-resource-type for its parent custom type because a resource check
step will now result in its parent being checked as the image plan. This
means we dont need to run an extra check to ensure the parent type is
checked. One problem that came about this was that when a user triggers
a new manually triggered build, this will result in the resources AND
its parent custom type forcing a check to happen. If they had 5
resources that all use the same custom type, it will force 5 checks on
the resources and then also 5 checks on the SAME custom type, compared
to before where it would have only forced 5 checks on the resources. So
in order to fix this extra performance burden, we made it so that there
will be a SkipInterval boolean on the CheckPlan and only if that is set
to true will the check be forced to skip its interval. And this will
only be set to true from fly check-resources or check-resource-types
where shallow is false. Every other case, like check webhooks or
manually triggered builds, that creates resource checks will not force a
check so SkipInterval will be false.
[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>