atc: remove atc.VersionedResourceTypes
Since we no longer use the version off of the VersionedResourceTypes
(because we always create a check whenever a custom type is used) inside
a build, we can remove this type entirely and replace all usages with
ResourceTypes. This commit replaces all the places where
VersionedResourceTypes was used within the ATC. We still need to convert
go-concourse and fly to no longer use it.

This might have some performance benefits because we no longer need to
query for a version for each resource type when we fetch the resource
types from the database.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>