B: disallow duplicate resource caches on same worker
with #6660, it was possible for a single worker to cache multiple copies
of the same resource_cache if it streamed in the same resource_cache
from multiple workers simultaneously.

this commit makes the worker_resource_caches unique on `(worker_name,
resource_cache_id)`, (ignoring the `worker_base_resource_type_id` for
purposes of uniqueness). in the event of a conflict, we indicate that
the returned WorkerResourceCache is invalid, in which case we silently
abort the InitializeResourceCache flow (much like how we silently abort
when we have a unique code violation when updating the volume, which
enforces there's only one volume for a single worker_resource_cache_id)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>