change default limits to non pointers
Having the field values be pointers in order to check if the value is
set or not by checking for a nil value is in some ways dangerous. If
someone forgets to do a nil check and the value is nil, this can easily
result in nil panics. Also with a config file, it will make it a lot
easier to check if the value was passed in or not through an enabled
boolean. So I switched over the field to be a value and an enabled
boolean to make it super clear whether or not a user intended to set the
limit or not. We will need some conversion between the flag and field
value, where only if the flag is passed in will we set the enabled
boolean to true.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>