Somehow copy-paste guardian's iptable rule to block host access
I expanded the integration test to also check that the container can
still reach the internet. I then decided to see what would happen if I
used guardian's iptable rule that mystified us and found that the test
stil passed! Both methods work but the nice thing about guardian's is
that you don't need to know the LAN IP address of the host. It just
works (TM).

I have a theory about why this works now, not 100% sure how to validate
it though.

When a process in a container reaches out to the hosts's LAN IP, the
path the packet takes is:

proc in container -> concourse0 bridge -> host process (gdn, baggageclaim, etc.)

When the process in a container is reaching out to an internet IP
address the path it takes is:

proc in container -> eth0 interface and directly out

Maybe the kernel is smart enough to figure this out and skips the
container bridge? This would explain why adding the reject rule directly
on the filter's INPUT chain doesn't break everything.

Guardian also adds two other rules before the reject rule:

-A w--input -i eth0 -j ACCEPT
-A w--input -m conntrack --ctstate RELATED,ESTABLISHED -j ACCEPT

These two never made sense to us. When we were investigating we
found no packets ever hit the first two rules, which makes sense. They
were part of a chain defined as:

-A INPUT -i w+ -j w--input

on the INPUT chain. This rule says, any packets going to an interface
that starts with "w" should go to the "w--input" chain. The first rule
in that chain:

-A w--input -i eth0 -j ACCEPT

makes no sense; packets on this chain are coming into "w+" interfaces,
it'll never match "eth0".

For the second rule, we get what it's meant to do, allow packets for
exisitng connections to pass through, we just don't know why you need it
here.

Assuming the above theory is correct, this single rule may be all we
need.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>