B: only update last_referenced in check step
not sure if this is required behaviour tbh, but a GC test started
failing because it expected the resource config that's created when we
create a resource cache to be immediately GC'able after the resource
cache is deleted.

previously, we only updated last_referenced when we
FindOrCreateResourceConfig (not ResourceCache). this commit brings back
that behaviour

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>