containerd: allow use of non-existent uids
Some images do not have an /etc/passwd file and need to be run with
non-existent uids.  Guardian accomodates for this by using the runc's
user.GetExecUserPath method.

Here, we can do the same. This method also takes
care of all the uid and gid parsing logic we were doing before as well as the
case of using the username root when the /etc/passwd file does not exist.

issue#6836

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>