B: add CheckPrototype API endpoint
nearly identical to CheckResourceType. one decision I made was to group
the audit log with resource auditing, which perhaps feels a little odd.
however, the prototype definition that we're checking is a resource
itself, so it makes sense

note that I'm not going to add a fly command yet - with the way
recursive checking is currently implemented, we'd need another API
endpoint to list prototypes (or get info on a particular prototype).
however, the recursive checking behaviour is going to be simplified in
600f965d02775f9212d2265677bf5ae228a62d30, so implementing this would be
pointless

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>