B: Containers reports active Tasks
this change was spurred by seeing checks error for a while after
redeploying our workers, e.g.:

```
run check: check: backend error: Exit status: 500, message: {"Type":"","Message":"task retrieval: no running task found: task 2adc1610-b027-4b0e-41e7-a7f9ebf74ee4 not found: not found","Handle":"","ProcessID":"","Binary":""}
```

the problem is that the ATC thinks the containers are still alive, so
does not invalidate the resource_config_check_session (until it expires
naturally, which can take up to an hour). digging into the worker, I
found that the Container actually *was* still alive - however, as the
error above suggests, the Task was not alive.

in containerd, a Container is just a metadata object that can contain a
Task. a Task is what we usually think of as a Container - a namespaced
process. so, what's happening when the worker restarts is the Container
persists, but the Task (i.e. the process) unsurprisingly gets
terminated. however, that Container doesn't get GC'd until the check
session expires because we report the Container as "alive", when it's in
fact not alive (if we consider the definition of Container used in the
rest of Concourse, i.e. an environment on which processes can be run)

so, with this change, the Garden `Containers` method only reports the
list of active Tasks, so when a Task dies, its corresponding Container
will eventually get GC'd.

note that we technically *could* start up a new Task on the Container if
we find the original task has been terminated, but this requires
additional bookkeeping to ensure we clean up the allocated network
namespaces for all of the tasks created for a Container. it's easier to
just treat a Container as a single-use thing and recreate them when
their process (task) dies

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>